Add perliminary DAG illustration
This diagram shows a simple repository with one local
branch ("topic") and another public branch ("master").

The commits that are not merged into the local "mater"
branch have a more pastel color than their local branch.
Also, the topic branch gets a different color from its
parent branch with a gradual transition.

The lower left hand side is an index explaining the
graphical elements. The title "DAG Widget", like the
index, is not also not part of the widget.